OK. So... TL;DR - fixed offset calculation of fields in a virtual object that has 64-bit field alignment.
After my fixes to clang were submitted there was a lot of churn with the MS ABI record layout stuff. Unfortunately, during these changes, my fixes were broken and the 3.0 branch was made.

... Bum ...

The good news is that the latest trunk code appears to work correctly but I'm not going to pull that in. I want to build from stable versions of clang/llvm each time from now on so I'll stick with the 3.0 branch until 3.1.

Fixed by partially restoring my old change.